requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-widget-tweaks==1.4.5
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
mysqlclient==1.4.4
Pillow==6.1.0
python-resize-image==1.1.19
pytz==2019.2
redis==3.3.8
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
